Digital Filter
Minimally functioning filter implementation. Filter coefficient math and window function still needs work.

+++ XANGELIX/Modules/EchoSampleProvider.cs
@@ -15,7 +15,7 @@ namespace XANGELIX.Modules {
 
 		public EchoSampleProvider() {
 			inputBuffer = new float[0];
-			echoCircleBuffer = new CircleBuffer(SampleRate);
+			echoCircleBuffer = new CircleBuffer((uint)SampleRate);
 
 			InputSampleProvider = new DCSampleProvider();
